# openapi.yaml
openapi: 3.1.0
info:
  title: Gatekeeper
  description: >
    API gateway on Cloudflare Workers with AWS IAM-style policy authorization.
    Fronts Cloudflare's cache purge API and S3-compatible R2 storage with
    per-key/credential policies, token-bucket rate limiting, request collapsing,
    and D1-backed analytics. The S3 proxy provides fine-grained object-level
    access control that R2's native tokens cannot — key-prefix restrictions,
    per-operation permissions, cross-bucket mixed policies.
  version: 3.0.0
  contact:
    name: Erfi Anugrah
servers:
  - url: https://gate.erfi.io
    description: Production

# ─── Security schemes ──────────────────────────────────────────────────────
security: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: >
        Gateway API key (prefix `gw_`). Issued via `POST /admin/keys`.
        Each key has an attached IAM policy that controls which actions,
        resources, and conditions it can access.

    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: >
        Shared admin secret for CLI and automation. Falls back when
        Cloudflare Access is not configured.

    CloudflareAccess:
      type: apiKey
      in: header
      name: Cf-Access-Jwt-Assertion
      description: >
        Cloudflare Access JWT. Injected automatically by Access for
        browser-based dashboard users. Also accepted via `CF_Authorization`
        cookie.

    S3SigV4Auth:
      type: apiKey
      in: header
      name: Authorization
      description: >
        AWS Signature Version 4 authentication. S3 clients (rclone, boto3,
        aws-cli) send this automatically. Credentials are issued via
        `POST /admin/s3/credentials` (access key prefix `GK`). Also
        supports presigned URL authentication via query parameters.

  # ─── Reusable schemas ──────────────────────────────────────────────────
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, errors]
      properties:
        success:
          type: boolean
          const: false
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string

    # --- Bulk operation results ---
    BulkItemResult:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [revoked, deleted, already_revoked, not_found]
    BulkResult:
      type: object
      required: [processed, results]
      properties:
        processed:
          type: integer
          description: Number of items processed.
        results:
          type: array
          items:
            $ref: '#/components/schemas/BulkItemResult'
    BulkDryRunItem:
      type: object
      required: [id, current_status, would_become]
      properties:
        id:
          type: string
        current_status:
          type: string
          enum: [active, revoked, expired, not_found]
        would_become:
          type: string
    BulkDryRunResult:
      type: object
      required: [dry_run, would_process, items]
      properties:
        dry_run:
          type: boolean
          const: true
        would_process:
          type: integer
          description: Number of items that would be processed.
        items:
          type: array
          items:
            $ref: '#/components/schemas/BulkDryRunItem'

    # --- Policy document ---
    PolicyDocument:
      type: object
      required: [version, statements]
      properties:
        version:
          type: string
          const: '2025-01-01'
        statements:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Statement'
...
    PurgeEvent:
      type: object
      required: [id, key_id, zone_id, purge_type, tokens, status, duration_ms, created_at]
      properties:
        id:
          type: integer
          description: Auto-incrementing event ID from D1.
        key_id:
          type: string
        zone_id:
          type: string
        purge_type:
          type: string
          enum: [host, tag, prefix, url, everything]
          description: >
            The type of purge operation. `url` = single-file purge by URL,
            `host`/`tag`/`prefix` = bulk purge by that dimension,
            `everything` = purge_everything.
        purge_target:
          type: ['string', 'null']
          description: >
            Human-readable summary of what was purged. For `url` purges,
            the comma-separated URLs (with `[Header:value]` for custom cache keys).
            For `host`/`tag`/`prefix`, the values. For `everything`, `"*"`.
            Truncated to 4 KB. `null` for events recorded before this field was added.
        tokens:
          type: integer
          description: >
            Rate-limit tokens consumed by this request. For `url` purges,
            equals the number of URLs. For bulk types, always 1.
        status:
          type: integer
        collapsed:
          type: ['string', 'null']
          description: >
            Request collapsing level: `"isolate"` (same isolate),
            `"do"` (Durable Object), or `null` if not collapsed.
        upstream_status:
          type: ['integer', 'null']
        duration_ms:
          type: integer
        response_detail:
          type: ['string', 'null']
          description: >
            Truncated upstream response for debugging. For purge requests,
            contains the CF API JSON `{success, errors, messages}`. Null when
            the request was collapsed or didn't reach upstream.
        created_by:
          type: ['string', 'null']
          description: >
            Identity of the caller — 'via API key' for token auth, email for Access SSO.
        created_at:
          type: integer
          description: Unix timestamp (ms).
    AnalyticsSummary:
      type: object
      required:
        - total_requests
        - total_urls_purged
        - by_status
        - by_purge_type
        - collapsed_count
        - avg_duration_ms
      properties:
        total_requests:
          type: integer
        total_urls_purged:
          type: integer
          description: Total number of individual URLs purged (single-file count, 1 per bulk request)
        by_status:
          type: object
          additionalProperties:
            type: integer
          examples:
            - { '200': 45, '429': 3 }
        by_purge_type:
          type: object
          additionalProperties:
            type: integer
          examples:
            - { 'host': 20, 'tag': 10, 'url': 15, 'everything': 3 }
        collapsed_count:
          type: integer
        avg_duration_ms:
          type: integer

    # --- S3 Credentials ---
    S3Credential:
      type: object
      required: [access_key_id, name, created_at, revoked, policy]
      properties:
        access_key_id:
          type: string
          description: Access key ID (prefix `GK`).
          examples: ['GK1A2B3C4D5E6F7890AB']
        secret_access_key:
          type: string
          description: >
            Secret access key. Only returned on creation — redacted in list/get.
        name:
          type: string
        created_at:
          type: integer
          description: Unix timestamp (ms).
        expires_at:
          type: ['integer', 'null']
          description: Unix timestamp (ms), or null if no expiration.
        revoked:
          type: integer
          enum: [0, 1]
          description: 0 = active, 1 = revoked.
        policy:
          type: string
          description: JSON-serialized PolicyDocument with S3 actions.
        created_by:
          type: ['string', 'null']
          description: Email of the creator (from Access JWT), or null.
    CreateS3CredentialRequest:
      type: object
      required: [name, policy]
      properties:
        name:
          type: string
        policy:
          $ref: '#/components/schemas/PolicyDocument'
          description: >
            Policy with S3 actions (e.g. `s3:GetObject`, `s3:PutObject`, `s3:*`)
            and S3 resources (e.g. `object:my-bucket/*`, `bucket:my-bucket`, `*`).
        expires_in_days:
          type: integer
          minimum: 1
        created_by:
          type: string

    # --- S3 Analytics ---
...
  parameters:
    ZoneIdQuery:
      name: zone_id
      in: query
      required: true
      schema:
        type: string
        pattern: '^[a-f0-9]{32}$'
      description: Cloudflare zone ID.
...
    AccessKeyIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: S3 credential access key ID (prefix `GK`).
      examples:
        - GK1A2B3C4D5E6F7890AB

# ─── Paths ─────────────────────────────────────────────────────────────────
paths:
...
  /admin/analytics/events:
    get:
      operationId: getAnalyticsEvents
      summary: Query purge event logs
      tags: [Analytics]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/ZoneIdQuery'
        - name: key_id
          in: query
          schema:
            type: string
          description: Filter events by API key ID.
        - name: since
          in: query
          schema:
            type: integer
          description: Start time (unix ms).
        - name: until
          in: query
          schema:
            type: integer
          description: End time (unix ms).
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 1000
          description: Max events to return.
      responses:
        '200':
          description: List of purge events.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/PurgeEvent'
        '400':
          description: Missing zone_id query parameter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: Analytics not configured (ANALYTICS_DB binding absent).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/analytics/summary:
    get:
      operationId: getAnalyticsSummary
      summary: Aggregated analytics
      tags: [Analytics]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/ZoneIdQuery'
        - name: key_id
          in: query
          schema:
            type: string
          description: Filter by API key ID.
        - name: since
          in: query
          schema:
            type: integer
          description: Start time (unix ms).
        - name: until
          in: query
          schema:
            type: integer
          description: End time (unix ms).
      responses:
        '200':
          description: Aggregated analytics for the zone.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    $ref: '#/components/schemas/AnalyticsSummary'
        '400':
          description: Missing zone_id query parameter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: Analytics not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ─── Admin: S3 Credentials ────────────────────────────────────────────────
  /admin/s3/credentials:
    post:
      operationId: createS3Credential
      summary: Create S3 credential
      description: >
        Create a new S3 credential with an attached IAM policy document.
        Returns the `access_key_id` (prefix `GK`) and `secret_access_key`.
        The secret is only shown on creation — it cannot be retrieved later.
        Use these credentials with any S3-compatible client (rclone, boto3,
        aws-cli) pointed at `https://<gateway>/s3`.
      tags: [S3Credentials]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateS3CredentialRequest'
            examples:
              full-access:
                summary: Full S3 access to all buckets
                value:
                  name: dev-full-access
                  policy:
                    version: '2025-01-01'
                    statements:
                      - effect: allow
                        actions: ['s3:*']
                        resources: ['*']
              read-only-prefix:
                summary: Read-only access to a key prefix
                value:
                  name: cdn-reader
                  policy:
                    version: '2025-01-01'
                    statements:
                      - effect: allow
                        actions: ['s3:GetObject', 's3:ListBucket']
                        resources: ['object:my-bucket/public/*', 'bucket:my-bucket']
                        conditions:
                          - field: key.prefix
                            operator: starts_with
                            value: 'public/'
                  expires_in_days: 90
      responses:
        '200':
          description: Credential created. Secret is only returned here.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    type: object
                    required: [credential]
                    properties:
                      credential:
                        $ref: '#/components/schemas/S3Credential'
                        description: >
                          The created credential. `secret_access_key` is only
                          present in this response — store it securely.
        '400':
          description: Validation error (missing fields, invalid policy).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

    get:
      operationId: listS3Credentials
      summary: List S3 credentials
      description: >
        Returns all S3 credentials. Secrets are redacted. Filter by
        status to get only active or revoked credentials.
      tags: [S3Credentials]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, revoked]
          description: Filter by credential status.
      responses:
        '200':
          description: List of S3 credentials (secrets redacted).
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/S3Credential'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/s3/credentials/{id}:
    get:
      operationId: getS3Credential
      summary: Get S3 credential details
      description: >
        Returns credential metadata including the serialized policy.
        The secret is redacted.
      tags: [S3Credentials]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/AccessKeyIdPath'
      responses:
        '200':
          description: Credential details (secret redacted).
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    type: object
                    required: [credential]
                    properties:
                      credential:
                        $ref: '#/components/schemas/S3Credential'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Credential not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

    delete:
      operationId: revokeOrDeleteS3Credential
      summary: Revoke or permanently delete an S3 credential
      description: >
        Without `permanent`, soft-revokes the credential (sets `revoked = 1`).
        The credential is immediately rejected on subsequent S3 requests
        (cache TTL up to 60s). With `permanent=true`, permanently removes the
        credential row from the database. Analytics events are preserved.
      tags: [S3Credentials]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/AccessKeyIdPath'
        - name: permanent
          in: query
          required: false
          schema:
            type: boolean
          description: >
            When `true`, permanently deletes the credential row instead of soft-revoking.
      responses:
        '200':
          description: Credential revoked or deleted.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    type: object
                    properties:
                      revoked:
                        type: boolean
                        description: Present when soft-revoking.
                      deleted:
                        type: boolean
                        description: Present when permanently deleting.
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Credential not found (or already revoked, when soft-revoking).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ─── Admin: Bulk S3 Credential Operations ────────────────────────────────
  /admin/s3/credentials/bulk-revoke:
    post:
      operationId: bulkRevokeS3Credentials
      summary: Bulk soft-revoke multiple S3 credentials
      description: >
        Soft-revoke multiple S3 credentials in one request. Requires `confirm_count` to match
        the number of IDs as a fat-finger guard. Set `dry_run: true` to preview without executing.
      tags: [S3 Credentials]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [access_key_ids, confirm_count]
              properties:
                access_key_ids:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 100
                  description: Array of access key IDs to revoke.
                confirm_count:
                  type: integer
                  description: Must exactly match the length of `access_key_ids`.
                dry_run:
                  type: boolean
                  default: false
                  description: Preview the operation without executing.
      responses:
        '200':
          description: Bulk revoke result or dry-run preview.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    oneOf:
                      - $ref: '#/components/schemas/BulkResult'
                      - $ref: '#/components/schemas/BulkDryRunResult'
        '400':
          description: Validation error (empty array, confirm_count mismatch, over 100 items).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/s3/credentials/bulk-delete:
    post:
      operationId: bulkDeleteS3Credentials
      summary: Bulk permanently delete multiple S3 credentials
      description: >
        Permanently remove multiple S3 credentials from the database. Works on credentials
        in any state. Requires `confirm_count` as a fat-finger guard. Set `dry_run: true` to preview.
        Analytics events referencing deleted credentials are preserved.
      tags: [S3 Credentials]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [access_key_ids, confirm_count]
              properties:
                access_key_ids:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 100
                  description: Array of access key IDs to delete.
                confirm_count:
                  type: integer
                  description: Must exactly match the length of `access_key_ids`.
                dry_run:
                  type: boolean
                  default: false
                  description: Preview the operation without executing.
      responses:
        '200':
          description: Bulk delete result or dry-run preview.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    oneOf:
                      - $ref: '#/components/schemas/BulkResult'
                      - $ref: '#/components/schemas/BulkDryRunResult'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ─── Admin: S3 Analytics ─────────────────────────────────────────────────
...